Feat/Refactor: 가게 정보 수정 페이지 API 구현 및 코드 리팩토링
